common expand_path
