+++ experiments/scenario_bowl2019/scenario_bowl2019/compare_approaches.py
@@ -47,10 +47,11 @@ def get_scores(args):
 def main(conf):
     logging.basicConfig(level=logging.INFO, format='%(asctime)s %(levelname)-7s %(funcName)-20s   : %(message)s')
 
+    embedding_file_names = sct.expand_path(conf['data_path'], conf['embedding_file_names'])
     approaches_to_train = {
         **{
             f"embeds: {file_name}": {'metric_learning_embedding_name': file_name}
-            for file_name in conf['embedding_file_names']
+            for file_name in embedding_file_names
         },
     }
     if conf['add_baselines']:
@@ -62,7 +63,7 @@ def main(conf):
         approaches_to_train.update({
             f"embeds: {file_name} and baseline": {
                 'metric_learning_embedding_name': file_name, 'use_handcrafted_features': True}
-            for file_name in conf['embedding_file_names']
+            for file_name in embedding_file_names
         })
 
     approaches_to_score = {
