+++ experiments/scenario_malware/scenario_malware/compare_approaches.py
@@ -46,21 +46,10 @@ def get_scores(args):
     return result
 
 
-def expand_path(data_path, wc_paths):
-    data_path = os.path.join(data_path, '')  # ensure `/` at the end
-
-    embedding_file_names = []
-    for path in wc_paths:
-        for n_path in glob(data_path + path):
-            embedding_file_names.append(n_path[len(data_path):])
-    logger.info(f'Found {len(embedding_file_names)} embedding files: [{embedding_file_names}]')
-    return embedding_file_names
-
-
 def main(conf):
     logging.basicConfig(level=logging.INFO, format='%(asctime)s %(levelname)-7s %(funcName)-20s   : %(message)s')
 
-    embedding_file_names = expand_path(conf['data_path'], conf['embedding_file_names'])
+    embedding_file_names = sct.expand_path(conf['data_path'], conf['embedding_file_names'])
     approaches_to_train = {
         **{
             f"embeds: {file_name}": {'metric_learning_embedding_name': file_name}
