+++ experiments/scenario_x5/scenario_x5/compare_approaches.py
@@ -1,12 +1,10 @@
 import logging
-import os
 from functools import partial, reduce
 from operator import iadd
 
 import numpy as np
 import pandas as pd
-from glob import glob
-from sklearn.metrics import roc_auc_score, make_scorer, accuracy_score
+from sklearn.metrics import make_scorer, accuracy_score
 
 import dltranz.scenario_cls_tools as sct
 from scenario_x5.const import (
@@ -63,21 +61,10 @@ def get_scores(args):
     return result
 
 
-def expand_path(data_path, wc_paths):
-    data_path = os.path.join(data_path, '')  # ensure `/` at the end
-
-    embedding_file_names = []
-    for path in wc_paths:
-        for n_path in glob(data_path + path):
-            embedding_file_names.append(n_path[len(data_path):])
-    logger.info(f'Found {len(embedding_file_names)} embedding files: [{embedding_file_names}]')
-    return embedding_file_names
-
-
 def main(conf):
     logging.basicConfig(level=logging.INFO, format='%(asctime)s %(levelname)-7s %(funcName)-20s   : %(message)s')
 
-    embedding_file_names = expand_path(conf['data_path'], conf['embedding_file_names'])
+    embedding_file_names = sct.expand_path(conf['data_path'], conf['embedding_file_names'])
     approaches_to_train = {
         **{
             f"embeds: {file_name}": {'metric_learning_embedding_name': file_name}
